pyproject.toml
[project]
name = "fastmcp-otel-langfuse"
version = "0.1.0"
description = "Example of distributed tracing with FastMCP, OpenTelemetry, and Langfuse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0rc1",
    "langfuse>=3.0.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation>=0.49b1",
    "opentelemetry-instrumentation-logging>=0.49b1",
    "opentelemetry-propagator-b3>=1.27.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.9.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["weather_assistant"]

[tool.ruff]
target-version = "py38"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.isort]
known-first-party = ["weather_assistant"]
